GRF-643 refactor classes organization and fix phpstan errors
