+++ src/Akeneo/Category/back/Application/Storage/UpdateCategoryEnrichedValues.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Akeneo\Category\Application\Storage;
+
+interface UpdateCategoryEnrichedValues
+{
+    /**
+     * @param array<string, string> $enrichedValuesByCode
+     */
+    public function execute(array $enrichedValuesByCode): void;
+}
