--- src/Akeneo/Category/back/Infrastructure/EventSubscriber/Cleaner/Sql/UpdateCategoryEnrichedValues.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace Akeneo\Category\Infrastructure\EventSubscriber\Cleaner\Sql;
-
-interface UpdateCategoryEnrichedValues
-{
-    public function execute(array $enrichedValuesByCode): void;
-}
